+++ netpyne/simFuncs.py
@@ -1413,7 +1413,8 @@ def gatherData (gatherLFP = True):
         if 'runTime' in sim.timingData:
             print('  Spikes: %i (%0.2f Hz)' % (sim.totalSpikes, sim.firingRate))
             if sim.cfg.printPopAvgRates and not sim.cfg.gatherOnlySimData:
-                sim.allSimData['popRates'] = sim.popAvgRates()
+            	trange = sim.cfg.printPopAvgRates if isinstance(sim.cfg.printPopAvgRates,list) else None
+                sim.allSimData['popRates'] = sim.popAvgRates(trange=trange)
             print('  Simulated time: %0.1f s; %i workers' % (sim.cfg.duration/1e3, sim.nhosts))
             print('  Run time: %0.2f s' % (sim.timingData['runTime']))