+++ CHANGES.md
@@ -1,3 +1,12 @@
+# Version 0.8.0
+
+- Added support for evolutionary algorithm optimization (via Inspyred) and usage example
+
+- cfg.popAvgRates now accepts a time range to calculate rates (e.g. to discard initial period)
+
+- Fixed bug initalizing batch 'mpi_bulletin' and batch tutorial example
+
+
 # Version 0.7.9
 
 - Extended metadata structure to interact with NetPyNE-UI